benchmark_bnd

compare two sets of breakend calls

benchmark_bnd(
  x,
  y,
  x.field = NA_character_,
  y.field = NA_character_,
  ignore.strand = FALSE,
  max.dist = 1000,
  genome = "hg19",
  std.only = TRUE,
  mask = system.file("extdata", "hg19.mask.rds", package = "gBenchmark"),
  verbose = FALSE
)

Arguments

x

character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (sv calls)

y

character vector containing a file path, GRangesList, gGraph, GRanges, or Junction object (ground truth)

x.field

(character) default NA

y.field

(character) default NA

ignore.strand

(logical) perform unsigned comparison of breakends? default FALSE

max.dist

(numeric) maximum distance between breakends to be considered a match, default 1 Kbp

genome

(character) one of hg19, hg3. default hg19

std.only

(logical) standard chromosomes only? default TRUE

mask

(character) path to coverage mask

verbose

(logical) message stuff? default FALSE

Details

given two sets of breakend calls x and y, with y being the 'ground truth' or 'gold standard' calculate precision, recall, and f1 score of x

x and y can be (signed) GRangesList, GRanges, gGraph (from which junctions/loose ends are extracted), or Junctions (or a file path containing any of these objects). in each of these cases, x.field and y.field do not have to be supplied.

in addition, x and y can be a segmentation output (such as GRanges) with some field containing copy number (supplied as x.field or y.field). using the copy number change points, a set of signed breakends are inferred.

if ignore.strand == TRUE, the orientation of breakends is ignored and two breakends in potentially opposite directions may be considered overlapping.

Examples

## compare a gGraph (x) with a set of junction calls provided as vcf (y)
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gg.rds", package = "gBenchmark"),
                                y = system.file("extdata", "svaba.subset.vcf", package = "gBenchmark"))
#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)
#> Warning: Each of the 2 combined objects has sequence levels not in the other:
#>   - in 'x': GL000191.1, GL000192.1, GL000193.1, GL000194.1, GL000195.1, GL000196.1, GL000197.1, GL000198.1, GL000199.1, GL000200.1, GL000201.1, GL000202.1, GL000203.1, GL000204.1, GL000205.1, GL000206.1, GL000207.1, GL000208.1, GL000209.1, GL000210.1, GL000211.1, GL000212.1, GL000213.1, GL000214.1, GL000215.1, GL000216.1, GL000217.1, GL000218.1, GL000219.1, GL000220.1, GL000221.1, GL000222.1, GL000223.1, GL000224.1, GL000225.1, GL000226.1, GL000227.1, GL000228.1, GL000229.1, GL000230.1, GL000231.1, GL000232.1, GL000233.1, GL000234.1, GL000235.1, GL000236.1, GL000237.1, GL000238.1, GL000239.1, GL000240.1, GL000241.1, GL000242.1, GL000243.1, GL000244.1, GL000245.1, GL000246.1, GL000247.1, GL000248.1, GL000249.1, MT, hs37d5
#>   - in 'y': M
#>   Make sure to always combine/compare objects based on the same reference
#>   genome (use suppressWarnings() to suppress this warning).
#> Warning: The 2 combined objects have no sequence levels in common. (Use
#>   suppressWarnings() to suppress this warning.)

## compare a gGraph (x) with a set of junction calls (provided as grl) (y)
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gg.rds", package = "gBenchmark"),
                                y = system.file("extdata", "grl.rds", package = "gBenchmark"))
#> Warning: Each of the 2 combined objects has sequence levels not in the other:
#>   - in 'x': GL000191.1, GL000192.1, GL000193.1, GL000194.1, GL000195.1, GL000196.1, GL000197.1, GL000198.1, GL000199.1, GL000200.1, GL000201.1, GL000202.1, GL000203.1, GL000204.1, GL000205.1, GL000206.1, GL000207.1, GL000208.1, GL000209.1, GL000210.1, GL000211.1, GL000212.1, GL000213.1, GL000214.1, GL000215.1, GL000216.1, GL000217.1, GL000218.1, GL000219.1, GL000220.1, GL000221.1, GL000222.1, GL000223.1, GL000224.1, GL000225.1, GL000226.1, GL000227.1, GL000228.1, GL000229.1, GL000230.1, GL000231.1, GL000232.1, GL000233.1, GL000234.1, GL000235.1, GL000236.1, GL000237.1, GL000238.1, GL000239.1, GL000240.1, GL000241.1, GL000242.1, GL000243.1, GL000244.1, GL000245.1, GL000246.1, GL000247.1, GL000248.1, GL000249.1, MT, hs37d5
#>   - in 'y': M
#>   Make sure to always combine/compare objects based on the same reference
#>   genome (use suppressWarnings() to suppress this warning).
#> Warning: Each of the 2 combined objects has sequence levels not in the other:
#>   - in 'x': GL000191.1, GL000192.1, GL000193.1, GL000194.1, GL000195.1, GL000196.1, GL000197.1, GL000198.1, GL000199.1, GL000200.1, GL000201.1, GL000202.1, GL000203.1, GL000204.1, GL000205.1, GL000206.1, GL000207.1, GL000208.1, GL000209.1, GL000210.1, GL000211.1, GL000212.1, GL000213.1, GL000214.1, GL000215.1, GL000216.1, GL000217.1, GL000218.1, GL000219.1, GL000220.1, GL000221.1, GL000222.1, GL000223.1, GL000224.1, GL000225.1, GL000226.1, GL000227.1, GL000228.1, GL000229.1, GL000230.1, GL000231.1, GL000232.1, GL000233.1, GL000234.1, GL000235.1, GL000236.1, GL000237.1, GL000238.1, GL000239.1, GL000240.1, GL000241.1, GL000242.1, GL000243.1, GL000244.1, GL000245.1, GL000246.1, GL000247.1, GL000248.1, GL000249.1, MT, hs37d5
#>   - in 'y': M
#>   Make sure to always combine/compare objects based on the same reference
#>   genome (use suppressWarnings() to suppress this warning).

## compare a GRanges (x) with a set of junction calls (provided as grl) (y) and ignore the strand
res = gBenchmark:::benchmark_bnd(x = system.file("extdata", "gr.rds", package = "gBenchmark"),
                                y = system.file("extdata", "grl.rds", package = "gBenchmark"),
                                ignore.strand = TRUE)
#> Warning: Each of the 2 combined objects has sequence levels not in the other:
#>   - in 'x': GL000191.1, GL000192.1, GL000193.1, GL000194.1, GL000195.1, GL000196.1, GL000197.1, GL000198.1, GL000199.1, GL000200.1, GL000201.1, GL000202.1, GL000203.1, GL000204.1, GL000205.1, GL000206.1, GL000207.1, GL000208.1, GL000209.1, GL000210.1, GL000211.1, GL000212.1, GL000213.1, GL000214.1, GL000215.1, GL000216.1, GL000217.1, GL000218.1, GL000219.1, GL000220.1, GL000221.1, GL000222.1, GL000223.1, GL000224.1, GL000225.1, GL000226.1, GL000227.1, GL000228.1, GL000229.1, GL000230.1, GL000231.1, GL000232.1, GL000233.1, GL000234.1, GL000235.1, GL000236.1, GL000237.1, GL000238.1, GL000239.1, GL000240.1, GL000241.1, GL000242.1, GL000243.1, GL000244.1, GL000245.1, GL000246.1, GL000247.1, GL000248.1, GL000249.1, MT, hs37d5
#>   - in 'y': M
#>   Make sure to always combine/compare objects based on the same reference
#>   genome (use suppressWarnings() to suppress this warning).