benchmark_cn

compare copy number (CN) predictions

benchmark_cn(
  x,
  y,
  x.field = "cn",
  y.field = "cn",
  allelic = FALSE,
  cnloh = FALSE,
  tile.width = 10000,
  min.width = 0,
  genome = "hg19",
  std.only = TRUE,
  mask = system.file("extdata", "hg19.mask.rds", package = "gBenchmark"),
  verbose = FALSE
)

Arguments

x

character vector containing a file path, data.table, GRanges, or gGraph

y

character vector containing a file path, data.table, GRanges, or gGraph

x.field

(character) field(s) in x containing CN data (default "cn")

y.field

(character) field(s) in y containing CN data (default "cn")

allelic

(logical) allele analysis? default FALSE

cnloh

(logical) cnloh analysis? default FALSE

tile.width

(numeric) tile width for averaging copy number estimates (default 10 Kbp)

min.width

(numeric) minimum width segment to retain (default 0, retains all segments)

genome

(character) one of hg19 or h38, default hg19

std.only

(character) only consider standard chromosomes? e.g. 1-22, X, Y. default TRUE

mask

(character) path to coverage mask

verbose

(logical) default FALSE

Value

data.table with benchmarking results

Details

checks consistency between total CN, allelic CN, or CNLOH predictions

inputs can be one of the follow objects, or a file containing one of the following objects: - bigwig - data.table coercible to GRanges - GRanges - gGnome gGraph

the "vanilla" analysis is comparison to total copy number and will return a data.table with Pearson correlation, Spearman correlation, and root mean square error (RMSE) computed over all genomic tiles

one variation of this analysis is comparison of major and minor allelic copy number. In this case, the input for x.field and y.field is expected to be a length two character vector, where the first element is the major allele CN field and the second element is the minor allele CN field. this will be run if allelic = TRUE.

another variation is comparison of copy-neutral LOH loci. In this case, the input for x.field and y.field is also a length two character vector (same as described above). this will be run if cnloh = TRUE. However, the output will be precision, recall, and F1 score of genomic tiles with cnLOH. Note that this comparison is non-symmetric and y is assumed to be the ground truth.

"unmappable" genomic tiles may also be masked by supplying a GRanges file path for the argument mask. coverage masks for hg19 and hg38 are included in extdata.

Examples

## compare the copy number information stored in a GRanges (x) with a data.table (y)
res = gBenchmark::benchmark_cn(x = system.file("extdata", "gr.rds", package = "gBenchmark"),
                               y = system.file("extdata", "dt.txt", package = "gBenchmark"),
                               x.field = "cn",
                               y.field = "cn",
                               tile.width = 1e3)
#> Warning: the standard deviation is zero
#> Warning: the standard deviation is zero

## compare the copy number information stored in a gGraph (x) to a GRanges (y)
res = gBenchmark::benchmark_cn(x = system.file("extdata", "gg.rds", package = "gBenchmark"),
                               y = system.file("extdata", "dt.txt", package = "gBenchmark"),
                               x.field = "cn",
                               y.field = "cn",
                               tile.width = 1e3)
#> Warning: the standard deviation is zero
#> Warning: the standard deviation is zero